+++ mail_2_rmq.py
@@ -741,11 +741,12 @@ def main():
     opt_xor_dict = {"-M": ["-C"], "-C": ["-M"]}
 
     # Process argument list from command line
-    args = gen_class.ArgParser(sys.argv, opt_val=opt_val_list, do_parse=True)
+    args = gen_class.ArgParser(sys.argv, opt_val=opt_val_list)
 
-    if not gen_libs.help_func(args, __version__, help_message)  \
-       and args.arg_require(opt_req=opt_req_list)               \
-       and args.arg_xor_dict(opt_xor_val=opt_xor_dict)          \
+    if args.arg_parse2()                                            \
+       and not gen_libs.help_func(args, __version__, help_message)  \
+       and args.arg_require(opt_req=opt_req_list)                   \
+       and args.arg_xor_dict(opt_xor_val=opt_xor_dict)              \
        and args.arg_dir_chk(dir_perms_chk=dir_perms_chk):
         run_program(args, func_dict)