+++ test/unit/mail_2_rmq/connect_process.py
@@ -17,7 +17,6 @@
 import sys
 import os
 import unittest
-import collections
 import mock
 
 # Local
@@ -36,9 +35,7 @@ class RQTest(object):
 
     Methods:
         __init__
-        create_connection
         publish_msg
-        change_channel
 
     """
 
@@ -52,26 +49,9 @@ class RQTest(object):
 
         """
 
-        self.exchange = "Test_Exchange"
         self.queue_name = "Test_Queue"
-        self.status = collections.namedtuple("RQ", "is_open")
-        self.channel = self.status(True)
-        self.conn_status = True
-        self.err_msg = ""
         self.pub_status = True
         self.msg = None
-
-    def create_connection(self):
-
-        """Method:  create_connection
-
-        Description:  Stub holder for create_connection method.
-
-        Arguments:
-
-        """
-
-        return self.conn_status, self.err_msg
 
     def publish_msg(self, msg):
 
@@ -86,18 +66,6 @@ class RQTest(object):
         self.msg = msg
 
         return self.pub_status
-
-    def change_channel(self, stat):
-
-        """Method:  change_channel
-
-        Description:  Change channel status.
-
-        Arguments:
-
-        """
-
-        self.channel = self.status(stat)
 
 
 class CfgTest(object):
@@ -141,10 +109,6 @@ class UnitTest(unittest.TestCase):
         test_true_publish
         test_error_queue
         test_non_error_queue
-        test_true_true_connect
-        test_false_false_connect
-        test_false_true_connect
-        test_true_false_connect
 
     """
 
@@ -321,82 +285,6 @@ class UnitTest(unittest.TestCase):
 
         self.assertFalse(
             mail_2_rmq.connect_process(self.rmq, mock_log, self.cfg, self.msg))
-
-    @mock.patch("mail_2_rmq.get_text")
-    @mock.patch("mail_2_rmq.gen_class.Logger")
-    def test_true_true_connect(self, mock_log, mock_msg):
-
-        """Function:  test_true_true_connect
-
-        Description:  Test connecting to RabbitMQ with true/true status.
-
-        Arguments:
-
-        """
-
-        mock_log.return_value = True
-        mock_msg.return_value = self.text
-
-        self.assertFalse(
-            mail_2_rmq.connect_process(self.rmq, mock_log, self.cfg, self.msg))
-
-    @mock.patch("mail_2_rmq.archive_email")
-    @mock.patch("mail_2_rmq.gen_class.Logger")
-    def test_false_false_connect(self, mock_log, mock_archive):
-
-        """Function:  test_false_false_connect
-
-        Description:  Test connecting to RabbitMQ with false/false status.
-
-        Arguments:
-
-        """
-
-        mock_log.return_value = True
-        mock_archive.return_value = True
-
-        self.rmq.conn_status = False
-        self.rmq.change_channel(False)
-        self.assertFalse(
-            mail_2_rmq.connect_process(self.rmq, mock_log, self.cfg, ""))
-
-    @mock.patch("mail_2_rmq.archive_email")
-    @mock.patch("mail_2_rmq.gen_class.Logger")
-    def test_false_true_connect(self, mock_log, mock_archive):
-
-        """Function:  test_false_true_connect
-
-        Description:  Test connecting to RabbitMQ with false/true status.
-
-        Arguments:
-
-        """
-
-        mock_log.return_value = True
-        mock_archive.return_value = True
-
-        self.rmq.conn_status = False
-        self.assertFalse(
-            mail_2_rmq.connect_process(self.rmq, mock_log, self.cfg, ""))
-
-    @mock.patch("mail_2_rmq.archive_email")
-    @mock.patch("mail_2_rmq.gen_class.Logger")
-    def test_true_false_connect(self, mock_log, mock_archive):
-
-        """Function:  test_true_false_connect
-
-        Description:  Test connecting to RabbitMQ with true/false status.
-
-        Arguments:
-
-        """
-
-        mock_log.return_value = True
-        mock_archive.return_value = True
-
-        self.rmq.change_channel(False)
-        self.assertFalse(
-            mail_2_rmq.connect_process(self.rmq, mock_log, self.cfg, ""))
 
 
 if __name__ == "__main__":