+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Pass filename along with attachment to RabbitMQ.
 
 ### Changed
+- connect_process: Inserted a file into a dictionary along with its filename.
 - connect_process: Removed the connection call, this moved to the calling functions - see below.
 - process_subj, process_from, process_file:  Added connection call and check to function.